production/agente-ia/requirements.txt
# FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# Google AI
google-generativeai>=0.8.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.116.0

# Database
supabase>=2.0.0
postgrest>=0.13.0

# Cache/Queue
redis[hiredis]>=5.0.0

# HTTP Client
httpx>=0.24.0
aiohttp>=3.9.0

# Scheduler
apscheduler>=3.10.0

# Utils
python-dotenv>=1.0.0
python-dateutil>=2.8.2
pytz>=2023.3
phonenumbers>=8.13.0

# Logging
structlog>=24.1.0
PyJWT

# PDF Processing
pymupdf>=1.24.0
email-validator>=2.0.0
